Few things: 	- Rename BooksController -> BookController 	- Rename BooksRoute -> BookRoute 	- Added babel package 	- Export static Class BookController 	- Export static Class IndexController
