+++ app/routes/IndexRoute.js
@@ -5,10 +5,10 @@
 * path: /
 ******************** */
 
-let express    = require('express');
-let Controller = require('../controllers/IndexController');
-let router     = express.Router();
+let express         = require('express');
+import IndexController from './../../app/controllers/IndexController';
+let router          = express.Router();
 
-router.get('/', Controller.index.get);
+// router.get('/', IndexController.index);
 
 module.exports = router;
